\name{check.components}
\alias{check.components}
\title{Interaction tool with the GWPCA glyph map}
\description{
The function interacts with the multivariate glyph plot of GWPCA loadings.
}
\usage{
check.components(ld,loc)
}

\arguments{
  \item{ld}{GWPCA loadings returned by \link{gwpca}}
  \item{loc}{a 2-column numeric array of GWPCA evaluation locations}
}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\seealso{\link{glyph.plot}}
\keyword{glyph plot, GWPCA}

