\name{LondonBorough}
\alias{LondonBorough}
\alias{londonborough}
\docType{data}
\title{London boroughs data}
\description{
Outline (SpatialPolygonsDataFrame) of London boroughs for the \link{LondonHP} data.
}
\usage{data(LondonBorough)}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\keyword{data}
\concept{London Boroughs}
