\name{print.gwrlcr}
\alias{print.gwrlcr}
\title{Print the output of the function \code{\link{gwr.lcr}}}
\description{
This function prints out the results from the function \link{gwr.lcr}.
}
\usage{
\method{print}{gwrlcr}(x, \dots)
}
\arguments{
  \item{x}{an object of class \dQuote{gwrlcr}, returned by the function \link{gwr.lcr}}
  \item{...}{arguments passed through (unused)}
}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\keyword{gwr.lcr, print}