\name{print.gwrm.pred}
\alias{print.gwrm.pred}
\title{Print the output of the function \code{\link{gwr.predict}}}
\description{
This function prints out the GWR results from the function \code{gwr.predict}
}
\usage{
\method{print}{gwrm.pred}(x, \dots)
}
\arguments{
  \item{x}{an object of class \dQuote{gwrm.pred}, returned by the function \link{gwr.predict}}
  \item{...}{arguments passed through (unused)}
}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\keyword{gwr, print}