\name{print.gwss}
\alias{print.gwss}
\title{Print the output of the function \code{\link{gwss}}}
\description{
This function prints out the results from \link{gwss}.
}
\usage{
\method{print}{gwss}(x, \dots)
}
\arguments{
  \item{x}{an object of class \dQuote{gwss}, returned by the function \link{gwss}}
  \item{...}{arguments passed through (unused)}
}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\keyword{gwss, print}