\name{EWOutline}
\alias{ewoutline}
\docType{data}
\title{Outline of England and Wales for data \link{EWHP}}
\description{
Outline (SpatialPolygonsDataFrame) of the England and Wales house price data \link{EWHP}.
}
\usage{data(EWOutline)}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\keyword{data}
\concept{England-Wales outline}
