\name{GWmodel-package}
\alias{GWmodel-package}
\alias{GWmodel}
\docType{package}
\title{
Geographically weighted models
}
\description{
In the R package GWmodel, we introduce techniques from a particular branch of spatial statistics, 
termed geographically weighted (GW) models. GW models suit situations when data are not described 
well by some global model, but where there are spatial regions where a suitably localised calibration 
provides a better description. GWmodel includes: GW summary statistics, GW principal components analysis,
GW discriminant analysis, Heteroskedastic GWR,GW regression, Mixed GWR,GW regression with a local ridge 
compensation, and GW regression for prediction; some of which are provided in basic and robust forms.
}
\details{
\tabular{ll}{
Package: \tab GWmodel\cr
Type: \tab Package\cr
Version: \tab 1.2-1\cr
Date: \tab 2013-07-09\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Binbin Lu, Paul Harris, Isabella Gollini, Martin Charlton, Chris Brunsdon

Maintainer: Binbin Lu <lubinbin220@gmail.com>
}
\note{
Acknowledgement: We gratefully acknowledge support from Science Foundation Ireland 
under the National Development Plan through the award of a Strategic Research Centre 
grant 07-SRC-I1168.
}

\keyword{ package }
