\name{print.ggwrm}
\alias{print.ggwrm}
\title{Print the output of the function \code{\link{gwr.generalised}}}
\description{
This function prints out the generalised GWR results from the function \code{\link{gwr.generalised}} with reference to the layout of the GWR3.0 software.
}
\usage{
\method{print}{ggwrm}(x, \dots)
}
\arguments{
  \item{x}{an object of class \dQuote{ggwrm}, returned by the function \link{gwr.generalised}}
  \item{...}{arguments passed through (unused)}
}
\author{Binbin Lu \email{binbinlu@whu.edu.cn}}
\references{
Charlton, M, Fotheringham, S, and Brunsdon, C (2007), GWR3.0.
}
\keyword{gwr, print}
