\name{print.gwrm}
\alias{print.gwrm}
\alias{print.summary.lm}
\title{Print the output of the function \code{\link{gwr.basic}}}
\description{
This function prints out the GWR results from the function \code{\link{gwr.basic}} with reference to the layout of the GWR3.0 software.
}
\usage{
\method{print}{gwrm}(x, \dots)
}
\arguments{
  \item{x}{an object of class \dQuote{gwrm}, returned by the function \code{\link{gwr.basic}}}
  \item{...}{arguments passed through (unused)}
}
\author{Binbin Lu \email{lubinbin220@gmail.com}}
\references{
Charlton, M, Fotheringham, S, and Brunsdon, C (2007), GWR3.0.
}
\keyword{gwr, print}
